Avoid modify config when passing source paths on command line
By adding back the source_paths property in Environment and adding a member
variable for it that is only used if it is set.

Do this since it will probably be possible to dump the configuration in the
future. But regardless, I think it is a good idea to not modify the config
dictionary. Should be able to rely on it being what is actually in the
configuration file.