Store environment in Check
Some checks will need access to the environment upon creation. An
alternative could be to lazy initialize some values that depend on
configuration and/or files in the project directory. I think delaying
creation until the environment is available is a better approach though.

Store check name as class variable in Check sub-classes. Fragile but do not
know if there is a better way to accomplish something like this in Python.