Tweak GPLv2, LGPLv2 and LGPLv2.1 license header texts
Line break more like in GPLv3/LGPLv3. Saves a bit of vertical space.

Remove old FSF address and replace it with the same URL that GPLv3/LGPLv3
refers to. (Address was changed a couple of years ago, some projects
updated and some projects changed to the URL. The URL is more future proof
and... who uses ordinary mail today...?)