Exclude generated source in build path when looking for source files
Common to create a build/ directory in the project root. If source is
generated in the build directory it will be found if 'source_paths' is not
set in .sork. Exclude it.

Meson requires build path and project path to be different. Other build
systems, e.g. CMake, does not though so have to check if they are the same
(build_path == os.path.curdir). Not much to do in that case other than
give up and include all the source. Another thing that this approach misses
is if there are multiple build directories. Not much to do in that case
either. User will have to set 'source_paths' in .sork to get reliable
behavior if build directories are used in any of these ways.