Disable showing score and storing of pylint data for later comparisons
Later versions of pylint (>= what is in Arch as of today at least)
apparently started to report score even if reports are disabled.

Also disable storing of persistent data because we are not interested in
comparing against older runs. We are only interested in being told where
the errors/warnings are.