Move use of ProgressPrinter from concurrent to commands
Moving ProgressPrinter to commands will make it easier to e.g. add a
verbose flag and to output name of file that is being processed. This is
not generic enough to have in concurrent.

A bit of repetition with try, catch, abort(), raise etc but will do
something about that at a later point if it is repeated more in e.g. other
commands.

Note that ProgressPrinter is now used from different threads. OK since it
was made thread safe in previous commit.