Fix bug for when no project source paths are specified
In .sork or on command line.

Project root passed on the command line is filtered out before
_find_file_paths() is called so that 'source_paths' in .sork is used when
only root is passed. If 'source_paths' is missing in .sork the source_paths
variable would be set to a string instead of an array of strings.

Dynamically typed languages for the fail!