Remove TODO about improving performance when searching for source files
Originally tested with Python 3.5. glob performance was improved in 3.6 .
- os.scandir() was added in 3.5: https://www.python.org/dev/peps/pep-0471
- glob started using it in 3.6: https://bugs.python.org/issue25596

Retesting searching for files in the Chromium source tree (~100000 source
files) with Python 3.6 results in 78s with a cold disk cache and 25s with
a warm cache. With 3.5 it took 2-3min regardless of warm/cold cache.

Searching for source files in the Linux source tree (~44000 source files)
the numbers are 10s with a cold cache and 3s with a warm cache.

All numbers are for source trees on a spinning and encrypted disk.

Most source trees are probably smaller than Chromium and Linux so I do not
think it is worth optimizing this at this point.