Add --verbose flag
For now only make it so that source files that are analyzed/checked are
printed on a single line above the progress string when Sork starts to
process them.

Causes pylint to warn about too many attributes in ProgressPrinter.
Cannot think of a nice way to split ProgressPrinter that results in fewer
attributes and does not unnecessarily complicate things. 7 (the default) is
a bit low, perhaps... might as well increase value to 10 globally instead
of only ignoring it for ProgressPrinter.