Ignore warnings in dependency include paths
By appending their include paths to C_INCLUDE_PATH and CPLUS_INCLUDE_PATH.
Both GCC and Clang use these variables to let the user specify additional
paths that should be seen as system include paths, for which warnings are
ignored. User is normally only interested in warnings and errors in own
code, not in external dependencies.

This also makes it so that clang-tidy ignores these headers.

Could be argued that users can specify these paths with -isystem. But that
is not always possible (e.g. pkg-config files use -I). And even if it is
possible users often overlook -isystem and use -I instead. This commit
makes it so Sork does the right thing (tm) for them.