Fix check target on Fedora
Fedora appends a -3 suffix to its Python 3 tools. Never seen pycodestyle3,
but check for it as well before checking for pycodestyle or pep8.

For some reason pycodestyle-3 in Fedora still uses the old [pep8] section.
Add it back but keep the new one (since pycodestyle on Arch emits warnings
if only [pep8] is found... *sigh*).

Should work on Arch, Debian and Fedora now.