Remove checks.NAMES
And just use [] as the default for checks to perform in config. Has the
same meaning as listing all checks explicitly.

Main reason for removing is that it causes some recursive import problems
in following commits when more stuff is imported in a lot of places for
type annotations. But, it makes sense without it. Do not really want a
dependency in config.py on checks. Should have a way to "register"
available config sections and variables from different modules. Will
implement this at some later point.