Verify configuration with JSON schema
Not sure how to handle jsonschema and mypy. Addint site-packages path
to MYPYPATH is no good (causes mypy to analyze every package in it and a
lot of them are not mypy friendly). Just ignore it for now. (mypy does
not look in site-packages by default so jsonschema is not found, hence
the --ignore-missing-imports argument.)