Remove use of JSON schemas
Added in 0c73e2a8425396d4f2f5fc9348f5101c394ea919 .

Initially noticed when writing tests that there was no way to easily verify
that default configuration was required to have all properties without a
lot of extra typing and/or repetition. Started looking into generating
schemas from default configuration (and having tuples for values that can
have different types). This started to be convoluted (no type safety and a
lot of magic with tuples). Decided to remove use of JSON schema and just
write config.Schema instead. Probably roughly the same amount of code but
more readable and understandable, I think.

As a bonus there is now no repetition of key names in dictionaries required
to define a configuration (previously they were in default dict and schema
dict). Less code is required to define default values/schema now as well.

There is also no longer any need to rely on an external dependency
(jsonschema) for validation. --ignore-missing-imports that was added to
mypy (see commit mentioned above) can therefore be removed.

The amount of test code increased a bit since we previously relied on
jsonschema doing the right thing (tm). If code were to be added for
automatically generating the JSON schema this would have to be tested as
well though. Also perhaps a bit bad to rely on jsonschema working 100%
without testing in Sork itself.