Extract functions for finding project and build path to separate module
And add tests for them.

Remove BuildPathFinder since it was mainly created in
5ff9285ed156b5b409e605b19a0e6363fd04e227 for grouping the functionality
of finding the build path together. Now when there is a separate module
for finding paths it is no longer needed for that reason. There was
really no other benefit of having it.

Rename some variables in environment.py from "paths" to something else
to not conflict with the module name.