Add a PopenMock class with helper method for patching
Will be used when testing Meson. Would like to test with actual file data
and use of actual executables but turned out to be hard (file data is not
simple text files).

Make sure going forward that mocking of Popen is not overused. Should still
prefer using actual commands even when testing.