Add arguments module
Makes main.py smaller and will make it easier (if desired) to test commands
etc. that deals with arguments. Will also try to move the 'source_paths'
logic here in following commits.

Have to set parser_class to add_subparsers() since it defaults to type
of parent. Do not want that. Want a regular argparse.ArgumentParser.