Move 'append to PATH like env variable' function to separate module
And add tests for it.

The 'environment' module name is available for env related things after
Environment was renamed to Project in 44625b723183 .

Also move handling of paths to append being empty to append function from
project._create_environment(). Makes sense (should be handled for all path
like env variables) and it is also easier to verify that it works correctly
in tests.