Remove imports in __init__.py of commands package
Commands are only imported in one place. Will add a constant in following
commit that would have to be added to __init__.py as well. Not worth it.

Actually want to move away from importing in __init__.py as much as
possible, but fix other packages at some later point.