Remove ArgumentParser subclass
Added in 22a8fb9531 when arguments module was extracted from main.py .

Not really needed though and just complicates things (e.g. the need for the
add_subparsers() parser_class arg, having to create an instance of it
in multiple places (main and tests) before parsing).

Move creation of parser and actual parsing to arguments.parse().