Add a CheckTestCase.create_check() method
Instead of creating check in setUp(). Will add possibility to have
different configuration, dependencies and compilation databases for
different tests in following commit. Means we cannot create check until we
are in test_* method.

A bit ugly to have Project as instance variable and add asserts to create
methods to check that they are called in correct order. :/