Add basic tests for ClangTidyCheck
Also modify check to return None instead of '' if there is no output from
clang-tidy at the same time. Does not matter for other callers at the
moment (may change in the future though and None is more correct, return
value is Optional) but matters when testing since we want to use
assertIsNone().