Nuke CheckTestCase
Hid away check creation in an ugly non-obvious manner. Saving of Project as
an instance variable and the asserts for it also prevented a test from
having multiple projects. Will probably require that in following
ClangTidyChecks.

Maybe it makes sense to have SourceFile creation in a common place, but not
worth it at the moment, I think.

TestCaseWithTmpDir.tmp_path() is modified to only prepend path to temporary
path if it is not already prepended. This allows for creating a tmp file
using e.g. Project.project_path, even if it already is an absolute path to
the project in the temporary directory, without having to have any special
handling for this where TestCaseWithTmpDir.create_tmp_file() is called.