Fix header filter bug in ClangTidyCheck with Python 3.7
The re module changed behavior in Python 3.7 compared to 3.6 which caused
most of ClangTidyHeaderFilterTestCase to fail.

E.g. the re.sub() statement in _header_filter_override() with
header_filter='^src/' and path='..' results in '^../src/' in 3.6. In 3.7
it results in '^../src/../'. That is, it is like 3.7 does a substitution
on a trailing empty string while 3.6 does not.

Changing '*' (match 0 or more) to '+' (match 1 or more) makes it so all
tests pass on both 3.6 and 3.7.