Move path related things from project module to paths module
Have to rename a couple of "paths" variables in source module now that
the paths module is imported in it.

Constant for normalized project path is made public so it can be used
directly. Replace function calls for normalizing project path with constant
use. It was private when path normalization was in project since it started
out as an implementation detail in Project.normalize_paths().