Fix searching for source with only project path specified
Should use paths from .sork or project path if "source_paths" is not set in
.sork .

The result of normalizing paths will be empty if only project path is
passed due to it being filtered out. Need to filter it to fallback to
"source_paths" in .sork as it is now. I think this has worked before but it
did not any longer.

The tests in SourceFindingTestCase() could be made a bit more readable. Do
not like that a comment is required when testing "source_paths" is not set
and that there is nothing when testing this case and "source_paths" is set.
Not obvious what is going on. Fix at some later point.