Get rid of final use of abc.abstractmethod in Check
Got rid of abc.abstractmethod for Command in
a95f2e06a3e30820cabba2e5f27cbbd3398f6fc5 . Tried to get rid of it in Check
then as well but was not happy with the result so scrapped it. This time
it went better. :)

Turn Check into a pod with a "run" callable and use closures etc instead of
subclasses for "storing" check specific information. More "pythonic", I
think.