Configure mypy to not allow untyped function definitions
And fix all violations.

Many of them are "-> None" missing for functions that do not return
anything. Liked that it was this way (nothing implied None). There is no way
to configure mypy to allow this when disallow_untyped_defs is set to True
though. So have to sprinkle "-> None" everywhere.

mypy also seems to be confused by unittest.TestCase.assertRaises() with
SystemError as exception in ArgumentsTestCase. Tests work as they should so
just annotate with "# type: ignore" until it is fixed.

mypy also fails on msg passed to unittest.TestCase.assertRaisesRegex() in
CompilationDatabaseTestCase. Not that important so just remove it.