Add line number to license header mismatch error
It should always be the first line in the source file, so hardcode a 1.

Probably does not make much of a difference, but if an editor e.g. just
leaves marker where it previously was if the user clicks the error message,
this will make it go to the top of the file where the license header is.