Do not allow project and author in license header to start with $ or space
And must contain at least a single character (same as before).

This prevents user from forgetting to replace $author and $project if copy
pasting from the template in the error message printed when there is a
license mismatch. (Not allowed to start with $.)

Must not start with space makes it so it is not possible to leave it out
completely or accidentally use " Foo" instead of "Foo".