Add SPDX license identifiers to license header lines
Apparently how to specify "or later" for the GPL family of licenses changed
from appending a "+" to "-or-later" sometime in December of 2017. See:

- https://spdx.org/ids-how
- https://www.gnu.org/licenses/identify-licenses-clearly.html
- spdx/license-list-XML@30cfeabd7a39206bfd22

Simply appending a plus was much shorter, do not understand why they had
to deprecate that. All other licenses still use the "+" format. *sigh* Oh
well. Follow suit and do not care about the identifiers that have been
deprecated for more than a year.

The Mozilla Public License Version 2.0 has "or later" built in to the
license itself. So nothing extra is required.

Does not look like Apache 2 has a "or later" built in but have not
investigated thoroughly. Not that interested in it personally.

Considered adding an option for letting user choose between only specifying
a SPDX identifier or not. But both all the GPL licenses and MPL vesrsion 2
recommends having the header text in each file. As does SPDX.