Use project path to get include guard prefix if no value in config
Makes it so that one does not have to have a .sork in the project root to
be able to use it.

Maybe a bit strange from a user perspective that behavior will change if
name of directory project is in is changed, but I think the downside of
this is out weighted but the benefits of having a better "out of the box"
experience.