Add source_paths_exclude configuration variable
For listing directories to be excluded when searching for source files and
set it to a list of common places where to put external projects by
default.

Could complicate this by e.g. listing git submodule paths and asking Meson
for what the subprojects path is set to, but decided to keep it simple.
Will not work out of the box if other locations are used, but should be
good enough for many use cases.